requirements.txt
# Core dependencies for Vehicle Submersion Detection System

# Deep Learning and Computer Vision
ultralytics>=8.0.0
opencv-python>=4.8.0
torch>=2.0.0
torchvision>=0.15.0

# Web Application
streamlit>=1.28.0

# Image Processing
Pillow>=10.0.0
numpy>=1.24.0

# Data Processing
pandas>=2.0.0

# Utilities
pathlib
